switch from tabs to spaces
Why the change of heart?
I did like the idea of using tabs instead of spaces because it allows
people to configure indentation width to their linking.
However, there a two problems.
One, to make formatting consistent you either have to stick with the
default tabwidth of 8 (which sucks) or you have to change the tabwidth
through all your systems to make it consistent. This is easily possible,
but then some programs expect a tabwidth of 8 (or more precisely a
tabstop of 8) and then break. Or you are ssh'ing on a system with a
tabstop of 8 while you have 4 and ls is just broken.
But realizing the other problem drove me to this descision: If everyone
can choose their indentation with, there is no way to get the 80char
convention right for everyone. Because this convention (or even 120char)
makes editing in small, tiled windows a lot nicer I prever having a
consistent 80char line length over using tabs.