remove dotbot
As i see it, dotbot doesn't net me anything: I am not using it's
plugins, so it gives me dependencies (even if only a git submodule) and
symlinks (the worse offender here) for no benefits over a pure git repo
(with detached git-dir) in return. In particular, the
adding-new-files-then-replacing-them-with-symlinks procedure is somewhat
unpleasant for me. Bare git just allows me to add a file without moving
it around, and on a new clone (just clone bare first) it can tell me the
differences instead of leaving me to manually diff the files.

One feature of a git wrapper that I would like is templating, for those
cases where configs can't be scripted suitably. However, all the
wrappers that do templating either impose too much on the repository
structure and workflow (e.g. chezmoi) or require obscure dependencies
that not even Arch has in its repos (yadm). Additionally, templated
files on prinicple do not allow me to directly edit and commit the
actual config file (the "result" of the template), which is a major
PITA. So instead I think I will live with per-host symlinks for those
cases where it cannot be avoided.