Remove dotbot dotfiles_local remnants
With dotbot I had a second, private repository with branches for every
host. That served both the purpose of having different files per host
and keeping sensitive files secret. It was however, a lot of work to
manage the branches, and from looking at the public dotfiles repo, you'd
have no idea where things of the private repo come in.
Now with the bare git repo, I move the first task (different files for
differen hosts) to the public repo, as should be, using scripting and/or
symlinks instead. I still need a private repo for sensitive files (not
credentials of course, they go in a password manager, but enumerations
like the SSH hosts), but now instead I deliberately symlink into this
repo, to make it obvious in the main (public) repo that here content
from the private repo is expected. That not only makes the public repo
more readable, it helps myself with not missing things and getting the
two repos out of sync.