have sway launch regular applications as transient services
Window managers don't usually make for good process managers, and sway
is no exception. Some applications are dbus-activated (which wofi can
do, because it just uses GIO which does it transparently), they are
already not directly launched by the window manager (but by dbus, or if
they are really cool, by dbus-activated systemd services). Gnome has
moved to a model where they put non-dbus activated programs into systemd
scopes after they launch them. I could have done the same (systemd-run
--scope is basically that), but I went with full transient services for
a reason: With scopes, processes are launched by the caller (e.g.
systemd-run which in turn is launched by sway) and then moved into the
scope, so they still inherit the environment of the caller, and their
stdout/stderr is connected to the caller. With a transient service, they
instead get forked off by the user manager, inheriting the explicitly
set environment and having their messages easier matched with the unit.

Of course, there are downsides to this approach because the semantics
don't fully match. For one, as I run wofi itself through systemd-run,
and wofi then forks off the actual application, I have to set
Type=forking unconditionally, and I cannot offer systemd a way to safely
determine the main PID, it has to guess. However, the actual
applications may be forking or they may not, the behaviour is not fixed
(though most of them probably won't fork). I could try to add an option
to wofis drun mode to create a transient service itself from the
application data, but that would be a major increase of code for the
mode (GIO does not directly support that), and then applications that do
fork themselves may break (though they might also behave weird right
now, as then with wofi a double-fork happens and systemd can't reliably
guess the main PID anymore). But if in the long run applications move to
dbus-activation this problem will solve itself anyway, and for the
moment this seems to work fine. There is just one funny side-effect of
this, I can go `systemctl restart` an application and ... wofi pops up.