move latex compilation into a subdirectory
Useful for easy removal of those twenty-something auxiliary files, and
easier to blacklist for sync clients (with as many extensions as a latex
build process has, collisions are practically guaranteed).
Unfortunately, TeXLive does not support a separate auxiliary directory,
only a directory for output in general. So, to get the rendered PDF out
I need some vim scripting...