set XDG_CURRENT_DESTKOP to sway
Well, set it for dbus and systemd-activated applications at least. Sway
both refuses to add this to their wayland session file as well as to
allow you to set environment variables from sway itself, so that can't
easily be done. At the same time, xdg-desktop-portal-wlr relies on this
being set to be found by xdg-desktop portal, all while it's very own
author claims that the variable should not be used. Sigh.

Since xdg-desktop-portal is launched via systemd for me, this should
work for now. Still, all systemd-run --scope launched apps inherit the
environment from sway, so I should find some solution for that. Locally
patching the wayland-session is probably the way to go, though I loathe
having to carry around that patch for all eternity, as upstream is not
going to add it...