remove the session-related envvars from systemd at logout
This is not really robust, when sway exits any other way than this
shortcut, unsetting is not going to happen. I could try to tie than in
through a service getting activated when the session goes away, maybe?
Tying the lifetime of environment variables in the activation
environment to the lifetime of a unit would be nice, but not sure if
that feature will be picked up upstream when a unit could also solve it.