migrate autostart to systemd services
For kdeconnect I also added a dbus service, even though I don't
bus-activate it (I want it to start in any case, so the phone can also
initiate a connection). This is only done because KDE ships it with a
dbus service that does _not_ launch a systemd service, but does it the
old way. Even though unlikely, I don't want to race a dbus-activation.