Revert "switch from tabs to spaces"
This reverts commit 06ebfe1a4fb23cdd1ddebf085c7a41a2af7dc721.

OK, so why switch back already? The reason is accessibility. The 80 or
120 character line width rule may work nice for me with my tiled
windows, but there may be reasons (other than style) for other people to
not have this fixed with and instead have a different indentation. Some
people might need larger indentation to easily discern the blocks and
accept a very wide line as a tradeoff. More than that, some people might
need smaller indentations as they may have to use large font sizes, and
still want to fit as much as possible onto the screen.
Whatever the case, I don't think anymore that the 80 character line
width rule necessarily is best for everyone.