switch from konsole to alacritty and termite
Konsole is still very nice, but QTs font rendering for OTF just looks
wrong. I initially wanted to switch to kitty, but kitty has no subpixel
rendering at all, making filigree fonts looking grainy and incomplete.
The stable version of alacritty first failed to run natively on sway at
all, and it is reported to focus on "line speed" over typing latency and
have other issues, so I went for termite first, as it appeared to be the
most lightweight terminal that runs natively on Wayland with acceptable
font rendering and plenty of supported escape codes.
I then tried the current git master of alacritty, which does work
natively on sway. Despite what is said about it, the percieved latency
from hitting a key to it showing up on the screen seems lower than what
termite has to offer. The dynamic_padding feature is really
aesthetically pleasing as well. There are some missing pieces though,
most importantly a scrollback search (or export, like kitty has), the
atrocious rendering of box drawing characters (apparently all
overlapping characters in general) and hyperlink escape code
recognition. Most of that is going to be fixed eventually, though the
devs are not planning on having alacritty take over rendering boxes, so
fancy prompts and htop are probably never going to look as nice as on
the other terminal emulators. Memory usage also seems to be very high
with alacritty, probably because of the GPU-accelerated scrollback.
I will try running alacritty as the default for now, but am prepared to
switch to termite should the issues become too annoying.