update desktop setup for the new monitor
Instead of a 23.5" 1080p monitor, my main one is now a 27" 4K one. This
improves the font rendering by a considerable amount, I can now have
text look both beautiful _and_ readable. The only downside is that
everything is huge now, since a non-integer scale factor is not
advisable (if there is _any_ content on the screen that is not scalable
from the get-go, e.g. legacy applications or bitmap-based UI in modern
applications, a non-integer scale factor will make these contents look
terrible).

As of this commit, Xwayland-based appliations are forcibly upscaled from
a simulated 1080p resolution. To fix this, I need to apply out-of-tree
patches for sway, wlroots and xwayland itself, and I need to split up
the sway configuration between the desktop setup and the notebook one,
since the patches introduce a global scale factor for Xwayland which
makes no sense on the notebook with its single 1080p screen.

Also, this monitor in theory supports 120Hz over DisplayPort 1.4. So
does my graphics card, but apparently only as long as no other monitor
is physically connected (even if no output is configured). For now, I
want to keep my secondary screens, and buying a new graphics card is
absolutely not an option right now (thanks, Bitcoin).