disable global subpixel font rendering
On the new monitor, grayscale rendering looks way better than subpixel
rendering (as the color fringes are more pronounced on it while there
are a lot more pixels per letter to use in the first place). Enabling it
only for the legacy screens _is_ possible, and already done via the sway
config, but few applications respect this (even though it seems to be
part of the base wayland protocol). Still, the grayscale renderer has
considerably improved even on low-dpi monitors, it now looks a lot more
like the subpixel rendering (still slightly fuzzier in exchange for
color friging, but the thickness and shapes of the letters seem to be
the exact same).