set a longer screen blanking timeout
The new monitor has very unfortunate behaviour in regards to screen
blanking (though it seems to be shared by other modern DisplayPort
monitors as well): When _resuming_ from standby it resets the
connection. This causes sway to move all workspaces over to a secondary
screen, resizing everything. This, in turn, changes the UI elements or
scrolling positions of many applications. Resizing and re-scrolling
everything back is not too big of an issue if done once or twice, but
doing it constantly becomes very annyoing. Setting the timeout to
several minutes prevents the resizing when I am just gone temporarily
from the machine.

I wonder if this behaviour is the result of an energy conservation
policy somewhere, which would be very ironic since it just caused me to
_not_ suspend my screen anymore.