sway: drive main monitor at 120Hz
I could not do this before, as the Radeon 570 can only output that much
on one monitor as long as no other monitors are connected; but a helpful
friend reminded me that my desktop also has an integrated GPU which I
can (with enough HDMI/DP to DVI adapters) use to drive the secondary
screens. It works, and now I can have both multiple monitors and 120Hz
refresh rate.

I disabled adaptive sync by default, as in desktop usage it can cause
some slight brightness flickering. This seems to happen when the refresh
rate jumps between the two extremes of 48 and 120Hz (whenever something
starts to move, basically), which only happens in regular desktop usage
and so far has never happened in a game. Thus for now I just turn it on
on-demand, it is fortunately toggleable at run-time.
This "flickering" seems to be common amongst "FreeSync" monitors, which
are the only monitors that are affected by the adaptive sync settings
(G-Sync monitors require Nvidia cards, which require using X11; Intel
does not yet have anything related to adaptive sync), so most sway users
should experience that. Maybe at some point sway will gain the ability
to set a lower bound to this setting.