ssh: switch to gcr-ssh-agent (gnome-keyring)
Now that I don't use a PGP key for SSH anymore, and that gnome-keyring
is just wrapping a regular ssh-agent (instead of reimplementing it and
thus not supporting every key type) there is no reason to use the PGP
agent anymore. At least not by default, some old systems still need my
PGP key of course, but this is then done in the ssh_config.