sway: remove screen placeholder
This 1280 offset was meant to enable me to add another screen
(physically already present) to my layout on demand, but it is causing a
lot of trouble for games and such. Considering that I haven't connected
that screen to my laptop ever and that I could just make another kanshi
profile if I do decide to move it from desktop to laptop, drop it.