bubbleenv: don't ro bind nix cache
Running the nix daemon from within the privileged nix env makes it
possible to have unprivileged environments install packages as well,
mediated via the daemon. However, it turns out that most of the
flake-stuff seems to happen locally, not on the daemon, and it needs to
write the cache. Sharing it RW defeats the isolation as far as I
understand it, so unfortunately this needs to be duplicated.